requirements_mac(no GPU).txt
# 金融术语标准化系统 - macOS CPU 依赖
# 适用于 macOS 环境，无需 GPU 支持
# 基于项目实际使用的包和版本

# ===== 核心 Web 框架 =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# ===== 数据验证 =====
pydantic==1.10.22
pydantic-settings==2.0.0

# ===== CORS 和文件上传支持 =====
python-multipart==0.0.5

# ===== 机器学习核心 (CPU 版本) =====
# PyTorch CPU 版本 - 适合 macOS Python 3.8
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
--extra-index-url https://download.pytorch.org/whl/cpu

# Transformers 生态 - Python 3.8 兼容版本
transformers==4.46.3
tokenizers==0.15.2
sentence-transformers==3.2.1
huggingface-hub==0.26.5
safetensors==0.4.5

# ===== 向量数据库 =====
pymilvus==2.5.14

# ===== 数据处理 =====
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.4

# ===== LangChain 生态系统 =====
# 使用兼容版本
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.42
langchain-openai==0.1.25
langchain-huggingface==0.1.2
langsmith==0.1.143

# ===== LLM 支持 =====
openai==1.97.1
# Ollama Python 客户端 (需要手动安装)
# ollama>=0.1.0

# ===== AWS 支持 (可选) =====
boto3==1.35.67
langchain-aws==0.2.4

# ===== 环境配置 =====
python-dotenv==1.0.1

# ===== HTTP 客户端 =====
requests==2.26.0
httpx==0.27.2
aiofiles==23.2.1

# ===== 科学计算 (macOS 优化) =====
scipy==1.10.1
scikit-learn==1.3.2

# ===== 异步支持 =====
anyio==4.6.2
sniffio==1.3.1

# ===== 开发和调试工具 =====
pytest==7.4.4
pytest-asyncio==0.21.2

# ===== 其他工具 =====
click==8.1.7
rich==13.9.4
typer==0.12.5

# ===== macOS 特定说明 =====
# 1. 此文件专为 macOS CPU 环境优化
# 2. 使用 CPU 版本的 PyTorch，避免 CUDA 依赖
# 3. 兼容 Python 3.8+
# 4. 适用于 Intel 和 Apple Silicon Mac
# 5. 建议在虚拟环境中安装

# ===== 安装说明 =====
# 推荐安装方式:
# 1. 创建虚拟环境: python3 -m venv venv
# 2. 激活环境: source venv/bin/activate
# 3. 升级 pip: pip install --upgrade pip
# 4. 安装依赖: pip install -r "requirements_mac(no GPU).txt"

# 或使用快速安装脚本:
# python3 start_macos.py
